op-dispute-mon: consolidate super agreement enricher tests
