+++ op-dispute-mon/mon/extract/super_agreement_enricher_test.go
@@ -467,3 +467,171 @@ func (s *stubSuperNodeClient) SuperRootAtTimestamp(_ context.Context, timestamp 
 		},
 	}, nil
 }
+
+// TestSuperNodeEndpointTracking verifies that all endpoint tracking fields are properly populated
+func TestSuperNodeEndpointTracking(t *testing.T) {
+	t.Run("TrackErrorsCorrectly", func(t *testing.T) {
+		validator, clients, _ := setupMultiSuperNodeTest(t, 3)
+		clients[0].outputErr = errors.New("error1")
+		clients[1].outputErr = errors.New("error2")
+		clients[2].superRoot = mockRootClaim
+		clients[2].derivedFromL1BlockNum = 100
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999, // Super root game type
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.NoError(t, err)
+
+		// Verify error tracking
+		require.Equal(t, 3, game.NodeEndpointTotalCount, "Should track total endpoints")
+		require.Equal(t, 2, game.NodeEndpointErrorCount, "Should track 2 errors")
+		require.Equal(t, 2, len(game.NodeEndpointErrors), "Should track 2 unique endpoint errors")
+		require.True(t, game.NodeEndpointErrors["client-0"], "Should track client-0 error")
+		require.True(t, game.NodeEndpointErrors["client-1"], "Should track client-1 error")
+	})
+
+	t.Run("TrackNotFoundCount", func(t *testing.T) {
+		validator, clients, _ := setupMultiSuperNodeTest(t, 3)
+		clients[0].notFound = true
+		clients[1].notFound = true
+		clients[2].superRoot = mockRootClaim
+		clients[2].derivedFromL1BlockNum = 100
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999,
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.NoError(t, err)
+
+		require.Equal(t, 3, game.NodeEndpointTotalCount)
+		require.Equal(t, 2, game.NodeEndpointNotFoundCount, "Should track 2 not found responses")
+		require.Equal(t, 0, game.NodeEndpointErrorCount, "Should have no errors")
+	})
+
+	t.Run("TrackSafeUnsafeCounts", func(t *testing.T) {
+		validator, clients, _ := setupMultiSuperNodeTest(t, 4)
+		// Two clients report safe (derivedFromL1BlockNum <= game.L1HeadNum)
+		clients[0].superRoot = mockRootClaim
+		clients[0].derivedFromL1BlockNum = 100 // Safe
+		clients[1].superRoot = mockRootClaim
+		clients[1].derivedFromL1BlockNum = 200 // Safe
+		// Two clients report unsafe (derivedFromL1BlockNum > game.L1HeadNum)
+		clients[2].superRoot = mockRootClaim
+		clients[2].derivedFromL1BlockNum = 201 // Unsafe
+		clients[3].superRoot = mockRootClaim
+		clients[3].derivedFromL1BlockNum = 300 // Unsafe
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999,
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		// This should result in disagreement due to mixed safety
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.NoError(t, err)
+
+		require.Equal(t, 4, game.NodeEndpointTotalCount)
+		require.Equal(t, 2, game.NodeEndpointSafeCount, "Should track 2 safe assessments")
+		require.Equal(t, 2, game.NodeEndpointUnsafeCount, "Should track 2 unsafe assessments")
+		require.True(t, game.HasMixedSafety(), "Should detect mixed safety")
+	})
+
+	t.Run("TrackDivergentSuperRoots", func(t *testing.T) {
+		validator, clients, _ := setupMultiSuperNodeTest(t, 3)
+		divergedRoot := common.HexToHash("0xdivergent")
+		clients[0].superRoot = mockRootClaim
+		clients[0].derivedFromL1BlockNum = 100
+		clients[1].superRoot = divergedRoot
+		clients[1].derivedFromL1BlockNum = 100
+		clients[2].superRoot = divergedRoot
+		clients[2].derivedFromL1BlockNum = 100
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999,
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.NoError(t, err)
+
+		require.True(t, game.NodeEndpointDifferentRoots, "Should flag divergent super roots")
+		require.False(t, game.AgreeWithClaim, "Should disagree when super roots diverge")
+	})
+
+	t.Run("TrackMixedAvailability", func(t *testing.T) {
+		validator, clients, _ := setupMultiSuperNodeTest(t, 3)
+		clients[0].notFound = true
+		clients[1].superRoot = mockRootClaim
+		clients[1].derivedFromL1BlockNum = 100
+		clients[2].superRoot = mockRootClaim
+		clients[2].derivedFromL1BlockNum = 100
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999,
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.NoError(t, err)
+
+		require.Equal(t, 3, game.NodeEndpointTotalCount)
+		require.Equal(t, 1, game.NodeEndpointNotFoundCount)
+		require.True(t, game.HasMixedAvailability(), "Should detect mixed availability")
+	})
+
+	t.Run("AllFieldsZeroWhenNoEndpoints", func(t *testing.T) {
+		logger := testlog.Logger(t, log.LvlInfo)
+		validator := NewSuperAgreementEnricher(logger, &stubOutputMetrics{}, []SuperRootProvider{}, clock.NewDeterministicClock(time.Unix(9824924, 499)))
+
+		game := &types.EnrichedGameData{
+			GameMetadata: challengerTypes.GameMetadata{
+				GameType: 999,
+			},
+			L1HeadNum:               200,
+			L2SequenceNumber:        0,
+			RootClaim:               mockRootClaim,
+			NodeEndpointErrors: make(map[string]bool),
+		}
+
+		err := validator.Enrich(context.Background(), rpcblock.Latest, nil, game)
+		require.ErrorIs(t, err, ErrSuperNodeRpcRequired)
+
+		// Verify all counts remain zero when no endpoints
+		require.Equal(t, 0, game.NodeEndpointTotalCount)
+		require.Equal(t, 0, game.NodeEndpointErrorCount)
+		require.Equal(t, 0, game.NodeEndpointNotFoundCount)
+		require.Equal(t, 0, game.NodeEndpointSafeCount)
+		require.Equal(t, 0, game.NodeEndpointUnsafeCount)
+		require.False(t, game.NodeEndpointDifferentRoots)
+	})
+}
